AR < 4 doesn't have SchemaMigration

+++ spec/database_rewinder_spec.rb
@@ -112,17 +112,19 @@ describe DatabaseRewinder do
     end
   end
 
-  describe '.clean_all' do
-    before do
-      ActiveRecord::SchemaMigration.create_table
-      ActiveRecord::SchemaMigration.create! version: '001'
-      Foo.create! name: 'foo1'
-      DatabaseRewinder.clean_all
-    end
-    after { ActiveRecord::SchemaMigration.drop_table }
-    it 'should clean except schema_migrations' do
-      Foo.count.should == 0
-      ActiveRecord::SchemaMigration.count.should == 1
+  if ActiveRecord::VERSION::STRING >= '4'
+    describe '.clean_all should not touch AR::SchemaMigration' do
+      before do
+        ActiveRecord::SchemaMigration.create_table
+        ActiveRecord::SchemaMigration.create! version: '001'
+        Foo.create! name: 'foo1'
+        DatabaseRewinder.clean_all
+      end
+      after { ActiveRecord::SchemaMigration.drop_table }
+      it 'should clean except schema_migrations' do
+        Foo.count.should == 0
+        ActiveRecord::SchemaMigration.count.should == 1
+      end
     end
   end
